+++ @xen-orchestra/lite/src/components/ui/UiTab.vue
@@ -9,7 +9,8 @@
 </template>
 
 <script lang="ts" setup>
-import { type ComputedRef, computed, inject } from "vue";
+import { IK_TAB_BAR_DISABLED } from "@/types/injection-keys";
+import { computed, inject } from "vue";
 
 withDefaults(
   defineProps<{
@@ -20,8 +21,8 @@ withDefaults(
   { tag: "span" }
 );
 
-const isTabBarDisabled = inject<ComputedRef<boolean>>(
-  "isTabBarDisabled",
+const isTabBarDisabled = inject(
+  IK_TAB_BAR_DISABLED,
   computed(() => false)
 );
 </script>