+++ @xen-orchestra/lite/src/views/pool/PoolRootView.vue
@@ -10,6 +10,11 @@
 <script lang="ts" setup>
 import PoolHeader from "@/components/pool/PoolHeader.vue";
 import PoolTabBar from "@/components/pool/PoolTabBar.vue";
+import { usePoolStore } from "@/stores/pool.store";
+import { usePageTitleStore } from "@/stores/page-title.store";
+
+const { pool } = usePoolStore().subscribe();
+usePageTitleStore().setObject(pool);
 </script>
 
 <style lang="postcss" scoped></style>